--- src/bin/__tests__/bin.test.ts
@@ -1,9 +1,0 @@
-import run from '../auto';
-
-test('throws error for unknown args', () => {
-  expect(
-    run({
-      command: 'foo'
-    })
-  ).rejects.toEqual(new Error(`idk what i'm doing.`));
-});
