+++ src/log-parse.ts
@@ -1,7 +1,7 @@
 import { ICommit } from 'parse-git';
 import { URL } from 'url';
 import join from 'url-join';
-import { ILogger } from './main';
+import { ILogger } from './GithubRelease';
 
 interface ICommitAuthor {
   name: string;
