+++ src/main.ts
@@ -1,351 +1,336 @@
+#!/usr/bin/env node
+
+import chalk from 'chalk';
+import cosmiconfig from 'cosmiconfig';
 import * as fs from 'fs';
-import { ICommit } from 'parse-git';
-import { inc, ReleaseType } from 'semver';
 import signale from 'signale';
 import { promisify } from 'util';
 
-import Github, { IGithubOptions, IPRInfo } from './git';
-import generateReleaseNotes, {
-  IExtendedCommit,
-  normalizeCommits
-} from './log-parse';
-import SEMVER, { calculateSemVerBump, ILabelMap } from './semver';
-import exec from './utils/exec-promise';
-import getGithubToken from './utils/github-token';
-import { dummyLog } from './utils/logger';
-import getConfigFromPackageJson from './utils/package-config';
-import postToSlack from './utils/slack';
+import parseArgs, { ArgsType } from './cli/args';
+import { IPRInfo } from './git';
+import GithubRelease, {
+  defaultLabels,
+  IGithubReleaseOptions
+} from './GithubRelease';
 
-export const defaultLabels = {
-  [SEMVER.major]: 'major',
-  [SEMVER.minor]: 'minor',
-  [SEMVER.patch]: 'patch',
-  'no-release': 'no-release',
-  release: 'release',
-  prerelease: 'prerelease'
-};
-
-export const defaultChangelogTitles = {
-  major: '💥  Breaking Change',
-  minor: '🚀  Enhancement',
-  patch: '🐛  Bug Fix',
-  internal: '🏠  Internal',
-  documentation: '📝  Documentation'
-};
-
-export interface ILogger {
-  log: signale.Signale<signale.DefaultMethods>;
-  verbose: signale.Signale<signale.DefaultMethods>;
-  veryVerbose: signale.Signale<signale.DefaultMethods>;
-}
-
-export interface IGithubReleaseOptions {
-  labels?: ILabelMap;
-  logger: ILogger;
-  jira?: string;
-  slack?: string;
-  githubApi?: string;
-  changelogTitles?: {
-    [label: string]: string;
-  };
-}
-
-export interface IOptionalGithubOptions {
-  owner?: string;
-  repo?: string;
-  baseUrl?: string;
-  token?: string;
-}
+import createLog from './utils/logger';
 
 const readFile = promisify(fs.readFile);
-const writeFile = promisify(fs.writeFile);
 
-export default class GithubRelease {
-  private readonly github: Promise<Github>;
-  private readonly userLabels: ILabelMap;
-  private readonly changelogTitles: { [label: string]: string };
-  private readonly logger: ILogger;
+const calcGreaterVersion = (
+  packageVersion: string,
+  monorepoVersion: string,
+  latestVersion: string
+) => {
+  const localVersion = monorepoVersion || packageVersion;
 
-  private readonly jira?: string;
-  private readonly slack?: string;
+  return localVersion > latestVersion ? localVersion : latestVersion;
+};
 
-  constructor(
-    options?: IOptionalGithubOptions,
-    releaseOptions: IGithubReleaseOptions = { logger: dummyLog() }
-  ) {
-    this.jira = releaseOptions.jira;
-    this.slack = releaseOptions.slack;
-    this.userLabels = releaseOptions.labels || {};
-    this.changelogTitles = releaseOptions.changelogTitles || {};
-    this.logger = releaseOptions.logger;
+async function getCurrentVersion(
+  prefixRelease: (release: string) => string,
+  lastRelease: string,
+  veryVerbose: signale.Signale<signale.DefaultMethods>
+) {
+  let lastVersion;
+  let packageVersion = '';
+  let monorepoVersion = '';
 
-    if (options && options.owner && options.repo && options.token) {
-      this.logger.verbose.info('Options contain repo information.');
+  if (fs.existsSync('lerna.json')) {
+    monorepoVersion = prefixRelease(
+      JSON.parse(await readFile('lerna.json', 'utf-8')).version
+    );
+  }
 
-      const args = {
-        owner: options.owner,
-        repo: options.repo,
-        logger: this.logger,
-        ...options
-      };
+  if (fs.existsSync('package.json')) {
+    packageVersion = prefixRelease(
+      JSON.parse(await readFile('package.json', 'utf-8')).version
+    );
+  }
 
-      if (releaseOptions && releaseOptions.githubApi) {
-        args.baseUrl = releaseOptions.githubApi;
-      }
+  if (lastRelease.match(/\d+\.\d+\.\d+/)) {
+    veryVerbose.info('Using latest release as previous version');
+    lastVersion = lastRelease;
+  } else if (monorepoVersion) {
+    veryVerbose.info('Using lerna.json as previous version');
+    lastVersion = monorepoVersion;
+  } else if (packageVersion) {
+    veryVerbose.info('Using package.json as previous version');
+    lastVersion = packageVersion;
+  } else {
+    veryVerbose.info('No previous release found, using 0.0.0 as the start.');
+    lastVersion = prefixRelease('0.0.0');
+  }
 
-      this.logger.verbose.info('Initializing Github API with:\n', args);
-      this.github = Promise.resolve(new Github(args));
-    } else {
-      this.logger.verbose.info('Getting repo information from package.json');
+  // This helps in situations where the latest release on github is wrong for some reason
+  // In this case we default to either the monorepo or package version
+  lastVersion = calcGreaterVersion(
+    packageVersion,
+    monorepoVersion,
+    lastRelease
+  );
 
-      this.github = getConfigFromPackageJson().then(async gOptions => {
-        const token = await getGithubToken();
+  return lastVersion;
+}
 
-        const finalOptions: IGithubReleaseOptions & IGithubOptions = {
-          ...options,
-          ...gOptions,
-          logger: this.logger,
-          token
-        };
+async function run(args: ArgsType) {
+  const logger = createLog(
+    args.very_verbose ? 'veryVerbose' : args.verbose ? 'verbose' : undefined
+  );
+  const { log, verbose, veryVerbose } = logger;
+  const explorer = cosmiconfig('auto');
+  const result = await explorer.search();
 
-        if (releaseOptions && releaseOptions.githubApi) {
-          finalOptions.baseUrl = releaseOptions.githubApi;
-        }
+  let rawConfig = {};
 
-        finalOptions.owner =
-          options && options.owner ? options.owner : gOptions.owner;
-        finalOptions.repo =
-          options && options.repo ? options.repo : gOptions.repo;
+  const prefixRelease = (release: string) =>
+    args.no_version_prefix || release.includes('v') ? release : `v${release}`;
 
-        this.logger.verbose.info(
-          'Initializing Github API with:\n',
-          finalOptions
+  if (result && result.config) {
+    rawConfig = result.config;
+  }
+
+  verbose.success('Loaded `auto-release` with config:', rawConfig);
+
+  Object.entries(args).forEach(([key, val]) => {
+    // tslint:disable-next-line
+    if (val === null) {
+      delete (args as any)[key];
+    }
+  });
+
+  const config: IGithubReleaseOptions = {
+    ...rawConfig,
+    ...args,
+    logger
+  };
+
+  let semVerLabels = defaultLabels;
+
+  if (config.labels) {
+    semVerLabels = {
+      ...defaultLabels,
+      ...config.labels
+    };
+  }
+
+  verbose.success('Using SEMVER labels:', '\n', semVerLabels);
+
+  const githubRelease = new GithubRelease(
+    {
+      owner: args.owner,
+      repo: args.repo
+    },
+    config
+  );
+
+  switch (args.command) {
+    case 'label': {
+      verbose.info("Using command: 'label'");
+      const labels = await githubRelease.getLabels(args.pr!);
+
+      console.log(labels.join('\n'));
+      break;
+    }
+    case 'pr-check': {
+      verbose.info(`Using command: 'pr-check' for '${args.url}'`);
+
+      args.target_url = args.url;
+      delete args.url;
+
+      let msg;
+
+      try {
+        const res = await githubRelease.getPullRequest(args.pr!);
+        args.sha = res.data.head.sha;
+
+        const labels = await githubRelease.getLabels(args.pr!);
+        const labelTexts = Object.values(semVerLabels);
+        const releaseTag = labels.find(l => l === 'release');
+        const noReleaseTag = labels.find(l => l === 'no-release');
+        const semverTag = labels.find(
+          l => labelTexts.includes(l) && l !== 'no-release' && l !== 'release'
         );
 
-        return new Github(finalOptions);
-      });
-    }
-  }
-
-  public async generateReleaseNotes(
-    from: string,
-    to = 'HEAD'
-  ): Promise<string> {
-    const client = await this.github;
-    const commits = await this.getCommits(from, to);
-    const project = await client.getProject();
-
-    await Promise.all(
-      commits.map(async commit => {
-        commit.packages = await client.changedPackages(commit.id);
-      })
-    );
-
-    return generateReleaseNotes(commits, {
-      owner: client.options.owner,
-      repo: client.options.repo,
-      baseUrl: project.data.html_url,
-      jira: this.jira,
-      logger: this.logger,
-      changelogTitles: {
-        ...defaultChangelogTitles,
-        ...this.changelogTitles
-      }
-    });
-  }
-
-  public async addToChangelog(
-    releaseNotes: string,
-    lastRelease: string,
-    noVersionPrefix = true,
-    message = 'Update CHANGELOG.md [skip ci]'
-  ) {
-    this.logger.verbose.info('Adding new changes to changelog.');
-
-    const version = await this.calcNextVersion(lastRelease);
-
-    this.logger.verbose.info('Calculated next version to be:', version);
-
-    const date = new Date().toDateString();
-    const prefixed = noVersionPrefix ? version : `v${version}`;
-
-    let newChangelog = `# ${prefixed} (${date})\n\n${releaseNotes}`;
-
-    if (fs.existsSync('CHANGELOG.md')) {
-      this.logger.verbose.info('Old changelog exists, prepending changes.');
-      const oldChangelog = await readFile('CHANGELOG.md', 'utf8');
-      newChangelog = `${newChangelog}\n\n---\n\n${oldChangelog}`;
-    }
-
-    await writeFile('CHANGELOG.md', newChangelog);
-    this.logger.verbose.info('Wrote new changelog to filesystem.');
-
-    await exec('git add CHANGELOG.md');
-    await exec(`git commit -m '${message}' --no-verify`);
-    this.logger.verbose.info('Commited new changelog.');
-  }
-
-  public async getCommits(
-    from: string,
-    to = 'HEAD'
-  ): Promise<IExtendedCommit[]> {
-    const client = await this.github;
-
-    this.logger.verbose.info(`Getting commits from ${from} to ${to}`);
-
-    const gitlog = await client.getGitLog(from, to);
-
-    this.logger.veryVerbose.info('Got gitlog:\n', gitlog);
-
-    const commits = await this.addLabels(gitlog);
-
-    this.logger.veryVerbose.info('Added labels to commits:\n', commits);
-
-    await Promise.all(
-      commits.map(async commit => {
-        let resolvedAuthors = [];
-
-        if (commit.pullRequest) {
-          const prCommits = await client.getCommitsForPR(
-            Number(commit.pullRequest.number)
-          );
-
-          if (!prCommits) {
-            return;
-          }
-
-          resolvedAuthors = await Promise.all(
-            prCommits.map(async prCommit => {
-              if (prCommit && prCommit.author) {
-                return client.getUserByUsername(prCommit.author.login);
-              }
-            })
-          );
-        } else if (commit.author.email) {
-          const author = await client.getUserByEmail(commit.author.email);
-          resolvedAuthors.push(author);
+        if (semverTag === undefined && !noReleaseTag) {
+          throw new Error('No semver label!');
         }
 
-        commit.authors = resolvedAuthors.map(author => ({
-          ...author,
-          username: author ? author.login : undefined
-        }));
+        log.success(`PR is using label: ${semverTag}`);
 
-        commit.authors.map(author => {
-          this.logger.veryVerbose.info(`Found author: ${author.username}`);
-        });
-      })
-    );
+        let description;
 
-    return commits;
-  }
-
-  public async publish(releaseNotes: string, tag: string) {
-    const client = await this.github;
-    return client.publish(releaseNotes, tag);
-  }
-
-  public async getLabels(pr: number) {
-    const client = await this.github;
-    return client.getLabels(pr);
-  }
-
-  public async createStatus(prInfo: IPRInfo) {
-    const client = await this.github;
-    return client.createStatus(prInfo);
-  }
-
-  public async getSha() {
-    const client = await this.github;
-    return client.getSha();
-  }
-
-  public async getLatestRelease(): Promise<string> {
-    const client = await this.github;
-    return client.getLatestRelease();
-  }
-
-  public async getPullRequest(pr: number) {
-    const client = await this.github;
-    return client.getPullRequest(pr);
-  }
-
-  public async createComment(message: string, pr: number, context = 'default') {
-    const client = await this.github;
-    return client.createComment(message, pr, context);
-  }
-
-  public async getSemverBump(
-    from: string,
-    to = 'HEAD',
-    onlyPublishWithReleaseLabel = false
-  ): Promise<SEMVER> {
-    const commits = await this.getCommits(from, to);
-    const labels = commits.map(commit => commit.labels);
-    const options = { onlyPublishWithReleaseLabel };
-    const versionLabels = {
-      ...defaultLabels,
-      ...this.userLabels
-    };
-
-    this.logger.verbose.info('Calculating SEMVER bump using:\n', {
-      labels,
-      versionLabels,
-      options
-    });
-
-    const result = calculateSemVerBump(labels, versionLabels, options);
-
-    this.logger.verbose.success('Calculated SEMVER bump:', result);
-
-    return result;
-  }
-
-  public async postToSlack(releaseNotes: string, tag: string) {
-    if (!this.slack) {
-      throw new Error('Slack url must be set to post a message to slack.');
-    }
-
-    const client = await this.github;
-    const project = await client.getProject();
-
-    this.logger.verbose.info('Posting release notes to slack.');
-
-    await postToSlack(releaseNotes, {
-      tag,
-      owner: client.options.owner,
-      repo: client.options.repo,
-      baseUrl: project.data.html_url,
-      slackUrl: this.slack
-    });
-
-    this.logger.verbose.info('Posted release notes to slack.');
-  }
-
-  public async calcNextVersion(lastTag: string) {
-    const bump = await this.getSemverBump(lastTag);
-    return inc(lastTag, bump as ReleaseType);
-  }
-
-  private async addLabels(commits: ICommit[]) {
-    if (!commits) {
-      return [];
-    }
-
-    const eCommits = normalizeCommits(commits);
-
-    await Promise.all(
-      eCommits.map(async commit => {
-        if (!commit.pullRequest) {
-          commit.labels = [];
+        if (noReleaseTag) {
+          description = 'PR will not create a release';
+        } else if (releaseTag) {
+          description = `PR will create release once merged - ${semverTag}`;
         } else {
-          commit.labels = await this.getLabels(
-            parseInt(commit.pullRequest.number, 10)
-          );
+          description = `CI - ${semverTag}`;
         }
-      })
-    );
 
-    return eCommits;
+        msg = {
+          description,
+          state: 'success'
+        };
+      } catch (error) {
+        msg = {
+          description: error.message,
+          state: 'error'
+        };
+      }
+
+      verbose.info('Posting comment to Github\n', msg);
+
+      if (!args.dry_run) {
+        await githubRelease.createStatus({
+          ...args,
+          ...msg
+        } as IPRInfo);
+        log.success('Posted status to Pull Request.');
+      } else {
+        verbose.info('`pr-check` dry run complete.');
+      }
+
+      verbose.success('Finished `pr-check` command');
+
+      break;
+    }
+    case 'pr': {
+      verbose.info("Using command: 'pr'");
+
+      if (!args.sha && args.pr) {
+        verbose.info('Getting commit SHA from PR.');
+        const res = await githubRelease.getPullRequest(args.pr);
+        args.sha = res.data.head.sha;
+      } else if (!args.sha) {
+        verbose.info('No PR found, getting commit SHA from HEAD.');
+        args.sha = await githubRelease.getSha();
+      }
+
+      verbose.info('Found PR SHA:', args.sha);
+
+      args.target_url = args.url;
+      delete args.url;
+
+      if (!args.dry_run) {
+        await githubRelease.createStatus(args as IPRInfo);
+      } else {
+        verbose.info('`pr` dry run complete.');
+      }
+
+      verbose.success('Finished `pr` command');
+
+      break;
+    }
+    case 'release': {
+      verbose.info("Using command: 'release'");
+
+      let lastRelease = await githubRelease.getLatestRelease();
+
+      // Find base commit or latest release to generate the changelog to HEAD (new tag)
+      veryVerbose.info(`Using ${lastRelease} as previous release.`);
+
+      if (lastRelease.match(/\d+\.\d+\.\d+/)) {
+        lastRelease = prefixRelease(lastRelease);
+      }
+
+      log.info('Last used release:', lastRelease);
+
+      const releaseNotes = await githubRelease.generateReleaseNotes(
+        lastRelease
+      );
+
+      log.info(`Using release notes:\n${releaseNotes}`);
+
+      const version =
+        args.use_version ||
+        (await getCurrentVersion(prefixRelease, lastRelease, veryVerbose));
+
+      if (!version) {
+        log.error('Could not calculate next version from last tag.');
+        return;
+      }
+
+      const prefixed = prefixRelease(version);
+      log.info(`Publishing ${prefixed} to Github.`);
+
+      if (!args.dry_run) {
+        await githubRelease.publish(releaseNotes, prefixed);
+
+        if (args.slack) {
+          log.info('Posting release to slack');
+          await githubRelease.postToSlack(releaseNotes, prefixed);
+        }
+      } else {
+        verbose.info('Release dry run complete.');
+      }
+
+      break;
+    }
+    case 'changelog': {
+      verbose.info("Using command: 'changelog'");
+
+      const lastRelease = args.from || (await githubRelease.getLatestRelease());
+      const releaseNotes = await githubRelease.generateReleaseNotes(
+        lastRelease,
+        args.to || undefined
+      );
+
+      log.info('New Release Notes\n', releaseNotes);
+
+      if (!args.dry_run) {
+        const currentVersion = await getCurrentVersion(
+          prefixRelease,
+          lastRelease,
+          veryVerbose
+        );
+        await githubRelease.addToChangelog(
+          releaseNotes,
+          currentVersion,
+          args.no_version_prefix,
+          args.message || undefined
+        );
+      } else {
+        verbose.info('`changelog` dry run complete.');
+      }
+
+      break;
+    }
+    case 'version': {
+      verbose.info("Using command: 'version'");
+
+      const lastRelease = await githubRelease.getLatestRelease();
+      const bump = await githubRelease.getSemverBump(
+        lastRelease,
+        undefined,
+        args.onlyPublishWithReleaseLabel
+      );
+
+      console.log(bump);
+
+      break;
+    }
+    case 'comment': {
+      verbose.info("Using command: 'comment'");
+
+      await githubRelease.createComment(
+        args.message!,
+        args.pr!,
+        args.context || undefined
+      );
+
+      log.success(`Commented on PR #${args.pr}`);
+      break;
+    }
+    default:
+      throw new Error(`idk what i'm doing.`);
+  }
+}
+
+export default async function main(args: ArgsType) {
+  try {
+    await run(args);
+  } catch (error) {
+    console.log(error);
   }
 }
